调整字面值表达式节点的位置
